style and file refactor
